--- web/rainmaker/dev-packages/egov-tradelicence-dev/src/ui-containers-local/RadioGroupContainer/index.css
@@ -1,14 +1,0 @@
-.radio-root {
-  margin-bottom: 0;
-  height: 30px;
-  color: #fe7a51;
-  width: 35px;
-}
-
-.radio-button-label {
-  margin-bottom: 0;
-  margin-right: 20px;
-  letter-spacing: 0.58px;
-  color: rgba(0, 0, 0, 0.87);
-  font-size: 16px;
-}
