+++ web/rainmaker/dev-packages/egov-ui-framework-core/src/ui-containers/index.js
@@ -33,8 +33,14 @@ const DownloadFileContainer = Loadable({
   loading: () => <Loading />
 });
 
+const RadioGroupContainer = Loadable({
+  loader: () => import("./RadioGroupContainer"),
+  loading: () => <Loading />
+});
+
 export {
   TextFieldContainer,
+  RadioGroupContainer,
   LabelContainer,
   MultiItem,
   SnackbarContainer,
